migrate bound join support to Typescript
