+++ src/stages/sage-bgp-stage-builder.ts
@@ -1,7 +1,7 @@
-/* file : sage-bgp-executor.js
+/* file : sage-bgp-executor.ts
 MIT License
 
-Copyright (c) 2018 Thomas Minier
+Copyright (c) 2018-2019 Thomas Minier
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -24,19 +24,24 @@ SOFTWARE.
 
 'use strict'
 
-const { stages } = require('sparql-engine')
-const boundJoin = require('../operators/bound-join.js')
+import { ExecutionContext, Graph, stages } from 'sparql-engine'
+import { Bindings } from 'sparql-engine/dist/rdf/bindings'
+import { PipelineStage } from 'sparql-engine/dist/engine/pipeline/pipeline-engine'
+import boundJoin from '../operators/bound-join'
+import { Algebra } from 'sparqljs'
 
 /**
- * Evaluate Basic Graph patterns using a Sage server
+ * Evaluate Basic Graph patterns using a Sage server.
+ * Contrary to the regular BGPStageBuilder,
+ * this subclass rely on the Bound Join algorithm to execute joins between BGPs.
+ * As the Sage server as a native support for UNION evaluation, it speeds up
+ * drastically the query execution by reducing the communication overhead.
  * @extends BGPExecutor
  * @author Thomas Minier
  * @author Corentin Marionneau
  */
-class SageBGPStageBuilder extends stages.BGPStageBuilder {
-  _buildIterator (source, graph, patterns, options, isJoinIdentity) {
-    return boundJoin(source, patterns, graph, options)
+export default class SageBGPStageBuilder extends stages.BGPStageBuilder {
+  _buildIterator (source: PipelineStage<Bindings>, graph: Graph, patterns: Algebra.TripleObject[], context: ExecutionContext): PipelineStage<Bindings> {
+    return boundJoin(source, patterns, graph, context)
   }
 }
-
-module.exports = SageBGPStageBuilder
